Add defaults.vim
I'm not sure that this exists at the same path on linux as it does on mac, and I'm not sure the `$VIMRUNTIME` var is working for linux. I'm pulling this into the project to be sure and might end up slowly removing parts of this default as things get implemented in vimrc.