Integrate system76 keyboard LEDs with ALE linting
The autocommands for this will only trigger if a system76 keyboard color
file is writeable. By default the file is owned by root, so you will
need to `chown` it. Please be aware that the file may be reverted back
to its original permissions during system upgrades.

Unfortunately, in this first iteration, the path `~/git/vim/keyboard` is
hard-coded. If you clone this repo to another location, you will have to
change the path.

set-error, set-warning, and restore all read in the file they are about
to write to in order to avoid rewriting the same value. This is better
for ssd longevity. This also takes into account that reading is 1.52
times faster than writing on this hard drive.

Tested on a [system76](https://system76.com/laptops) darter pro :)